Refactor AlertManager structure
