--- vitaes-webapp/src/Util.js
@@ -1,12 +1,0 @@
-
-
-
-
-
-
-
-
-
-
-
-
